Move license and SSO code into separate files. Fix broken default export
