+++ packages/back-end/src/init/index.ts
@@ -1,4 +1,4 @@
-import licenseInit from "enterprise";
+import { licenseInit } from "enterprise";
 import { logger } from "../util/logger";
 import mongoInit from "./mongo";
 import { queueInit } from "./queue";
